finished

+++ Swiatlomierz/Include/Src/BH1750.c
@@ -4,7 +4,7 @@
 float ReadLight(){
 
 	uint8_t datain[1]={0x20};
-	uint16_t tab[2] = {10,10};
+	uint8_t tab[2] = {10,10};
 	tw_init(TW_FREQ_400K,true);
 	if(tw_master_transmit(0x23, datain,1,false) != 0)
 	{
